Data types

+++ libraries/Sensor/Sensor.h
@@ -7,6 +7,7 @@
 
 #include <Arduino.h>
 #include <string.h>
+#include <math.h>
 
 class Sensor
 {
@@ -19,8 +20,8 @@ class Sensor
     float lo_thresh;
     float conv_coef;
     float conv_offset;
-    int pin;
-    int sensor_read;
+    unsigned short int pin;
+    unsigned short int sensor_read;
     float display_value;
     String display_name;
     // Methods
